resolved circular dependencies between tab scaffold and tab screens by introducing tab_navigation leaf module for shared navigation state and updating all importing widgets, services, and tests
