+++ test/test_provider_scope.dart
@@ -25,7 +25,7 @@ import 'package:lichess_mobile/src/network/aggregator.dart';
 import 'package:lichess_mobile/src/network/connectivity.dart';
 import 'package:lichess_mobile/src/network/http.dart';
 import 'package:lichess_mobile/src/network/socket.dart';
-import 'package:lichess_mobile/src/tab_scaffold.dart' show rootNavRouteStackObserver;
+import 'package:lichess_mobile/src/tab_navigation.dart' show rootNavRouteStackObserver;
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 import 'package:visibility_detector/visibility_detector.dart';
